Module wave::recognizer

the main entry point in the library.

All distance values are considered to be in millimeter

Structs

GestureRecognizer

The gesture recognizer. Is initially configured through parameters and gets fed measurements and time and predicts gestures based on its input.

RecognizerParams

Parameters for gesture recognition

RecognizerResult

A gesture prediction result.

Enums

Gesture

Represents a hand gesture

RecognizerStatus

The status of the gesture recognizer

Constants

HISTORY_SIZE

The history size.