Module wave::cbind

wave C bindings

Constants

HISTORY_SIZE

The history size.

RES_X

The horizontal sensor resolution.

RES_Y

The vertical sensor resolution.

Functions

coords_cartesian_from_spherical

Converts spherical to cartesian coordinates

coords_spherical_from_cartesian

Converts cartesian to spherical coordinates

gesture_recognizer_new

A new gesture recognizer with the given parameters.

gesture_recognizer_reset

Resets the gesture recognizer with the given parameters. Clears the history, ongoing predictions and resets the internal state.

gesture_recognizer_update

Updates the gesture recognizer with a new measurement and time.

recognizer_params_default

Default recognizer parameters, providing a good starting point for gesture recognition

recognizer_result_default

Default result, meaning no hand is found and no gestures are recognized. Used to initialize the result before passing it to the recognizer

sensor_measurement_invalid

Creates an invalid measurement ( distances are set to -1.0 )

sensor_params_default_vl53l5cx

Default sensor parameters for the ST VL53L5CX TOF-Sensor